Extract ECAlgorithm from ECDsa and ECDiffieHellman
