+++ src/libraries/System.Security.Cryptography/src/System/Security/Cryptography/ECDiffieHellman.Create.Android.cs
@@ -3,7 +3,7 @@
 
 namespace System.Security.Cryptography
 {
-    public abstract partial class ECDiffieHellman : AsymmetricAlgorithm
+    public abstract partial class ECDiffieHellman : ECAlgorithm
     {
         public static new partial ECDiffieHellman Create()
         {
