+++ src/libraries/System.Security.Cryptography/src/System/Security/Cryptography/ECDiffieHellman.Create.NotSupported.cs
@@ -3,7 +3,7 @@
 
 namespace System.Security.Cryptography
 {
-    public partial class ECDiffieHellman : AsymmetricAlgorithm
+    public partial class ECDiffieHellman : ECAlgorithm
     {
         public static new partial ECDiffieHellman Create()
         {
