+++ src/libraries/System.Security.Cryptography/src/System/Security/Cryptography/ECDsa.Create.Android.cs
@@ -8,7 +8,7 @@ using Microsoft.Win32.SafeHandles;
 
 namespace System.Security.Cryptography
 {
-    public partial class ECDsa : AsymmetricAlgorithm
+    public partial class ECDsa : ECAlgorithm
     {
         /// <summary>
         /// Creates an instance of the platform specific implementation of the cref="ECDsa" algorithm.
